+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
-## [Unreleased]
+## [8.0.0b2] - 2024-06-26
 
 ### Added
 
@@ -1503,7 +1503,8 @@ This release contains no changes except for the version number.
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/dipdup-io/dipdup/compare/8.0.0b1...HEAD
+[Unreleased]: https://github.com/dipdup-io/dipdup/compare/8.0.0b2...HEAD
+[8.0.0b2]: https://github.com/dipdup-io/dipdup/compare/8.0.0b1...8.0.0b2
 [8.0.0b1]: https://github.com/dipdup-io/dipdup/compare/8.0.0a1...8.0.0b1
 [7.5.7]: https://github.com/dipdup-io/dipdup/compare/7.5.6...7.5.7
 [7.5.6]: https://github.com/dipdup-io/dipdup/compare/7.5.5...7.5.6